+++ src/steps/intune/steps/applications/__tests__/index.test.ts
@@ -5,10 +5,10 @@ import {
 import { setupAzureRecording } from '../../../../../../test/recording';
 import { config } from '../../../../../../test/config';
 import { fetchDevices } from '../../devices';
-import { entities, relationships } from '../../../constants';
-import { isEqual } from 'lodash';
+import { entities } from '../../../constants';
 import { toArray } from '../../../../../utils/toArray';
 import { fetchDetectedApplications, fetchManagedApplications } from '..';
+import { groupBy, sortBy, last, uniq } from 'lodash';
 
 let recording: Recording;
 
@@ -19,7 +19,7 @@ afterEach(async () => {
 });
 
 describe('fetchManagedApplications', () => {
-  test('should make entities and relationships correctly', async () => {
+  it('should make entities and relationships correctly', async () => {
     recording = setupAzureRecording({
       directory: __dirname,
       name: 'fetchManagedApplications',
@@ -29,28 +29,25 @@ describe('fetchManagedApplications', () => {
     await fetchDevices(context);
     await fetchManagedApplications(context);
 
-    const managedApplicationEntities = context.jobState.collectedEntities.filter(
-      (e) => isEqual(e._class, toArray(entities.MANAGED_APPLICATION._class)),
+    const nonHostOrHostAgentEntities = context.jobState.collectedEntities.filter(
+      (e) =>
+        !toArray(e._class).includes('HostAgent') &&
+        !toArray(e._class).includes('Host'),
     );
     const deviceApplicationRelationships = context.jobState.collectedRelationships.filter(
-      (r) =>
-        relationships.MULTI_DEVICE_ASSIGNED_MANAGED_APPLICATION.map(
-          (c) => c._type,
-        ).includes(r._type),
+      (r) => r._type.includes(entities.MANAGED_APPLICATION._type),
     );
 
-    // Check that we have Managed Applications
-    expect(managedApplicationEntities.length).toBeGreaterThan(0);
-    expect(managedApplicationEntities).toMatchGraphObjectSchema({
-      _class: entities.MANAGED_APPLICATION._class,
-    });
-    expect(managedApplicationEntities).toMatchSnapshot(
-      'managedApplicationEntities',
-    );
+    // Check that we have no entities
+    expect(nonHostOrHostAgentEntities.length).toBe(0);
 
     // Check that we have DEVICE_ASSIGNED_MANAGED_APPLICATION relationships
     expect(deviceApplicationRelationships.length).toBeGreaterThan(0);
-    expect(deviceApplicationRelationships).toMatchDirectRelationshipSchema({});
+    deviceApplicationRelationships.forEach((r) =>
+      expect(r).toMatchObject({
+        _mapping: expect.any(Object),
+      }),
+    );
     expect(deviceApplicationRelationships).toMatchSnapshot(
       'deviceManagedApplicationRelationships',
     );
@@ -58,7 +55,7 @@ describe('fetchManagedApplications', () => {
 });
 
 describe('fetchDetectedApplications', () => {
-  test('should make entities and relationships correctly', async () => {
+  it('should make entities and relationships correctly', async () => {
     recording = setupAzureRecording({
       directory: __dirname,
       name: 'fetchDetectedApplications',
@@ -68,30 +65,42 @@ describe('fetchDetectedApplications', () => {
     await fetchDevices(context);
     await fetchDetectedApplications(context);
 
-    const detectedApplicationEntities = context.jobState.collectedEntities.filter(
-      (e) => isEqual(e._class, toArray(entities.DETECTED_APPLICATION._class)),
+    const nonHostOrHostAgentEntities = context.jobState.collectedEntities.filter(
+      (e) =>
+        !toArray(e._class).includes('HostAgent') &&
+        !toArray(e._class).includes('Host'),
     );
     const deviceApplicationRelationships = context.jobState.collectedRelationships.filter(
-      (r) =>
-        relationships.MULTI_DEVICE_HAS_DETECTED_APPLICATION.map(
-          (c) => c._type,
-        ).includes(r._type),
+      (r) => r._type.includes(entities.DETECTED_APPLICATION._type),
     );
 
-    // Check that we have Detected Applications
-    expect(detectedApplicationEntities.length).toBeGreaterThan(0);
-    expect(detectedApplicationEntities).toMatchGraphObjectSchema({
-      _class: entities.DETECTED_APPLICATION._class,
-    });
-    expect(detectedApplicationEntities).toMatchSnapshot(
-      'detectedApplicationEntities',
-    );
+    // Check that we have no entities
+    expect(nonHostOrHostAgentEntities.length).toBe(0);
 
-    // Check that we have DEVICE_ASSIGNED_DETECTED_APPLICATION relationships
+    // Check that we have DEVICE_ASSIGNED_MANAGED_APPLICATION relationships
     expect(deviceApplicationRelationships.length).toBeGreaterThan(0);
-    expect(deviceApplicationRelationships).toMatchDirectRelationshipSchema({});
+    deviceApplicationRelationships.forEach((r) =>
+      expect(r).toMatchObject({
+        _mapping: expect.any(Object),
+      }),
+    );
     expect(deviceApplicationRelationships).toMatchSnapshot(
       'deviceDetecctedApplicationRelationships',
     );
+
+    // Check that you can have multiple relationships with the same application based on version
+    const groupedRelationships = groupBy(
+      deviceApplicationRelationships,
+      '_mapping.targetEntity.name',
+    );
+    const appWithMultipleVersions = last(
+      sortBy(groupedRelationships),
+      (c) => c.length,
+    );
+    expect(appWithMultipleVersions.length).toBeGreaterThan(1);
+    // Check that all versions are unique
+    const versions = appWithMultipleVersions.map((app) => app.version);
+    expect(versions.length).toBeGreaterThan(0);
+    expect(versions.length).toEqual(uniq(versions).length);
   });
 });